Output signal is measured by ADC module and result of adc is mapped and a PWM signal is created with adc result on PIC18F4XK22
ADC module and PWM module is used together with PIC18F4XK22.
Output signal is measured by ADC module and result of adc is mapped and a PWM signal is created with adc result on PIC18F4XK22.
This project is written with C Language on MikroC PRO for PIC.